Replace bfabric.py by bfabric2.py
To ensure a smooth transition for new code, the existing bfabric2.py module is kept with reexports
of its old contents.

The old code was renamed to bfabric_legacy.py
