+++ bfabric/__init__.py
@@ -2,12 +2,9 @@ import importlib.metadata
 
 __version__ = importlib.metadata.version("bfabric")
 
-name = "bfabricPy"
-alias = "suds-py3"
+from bfabric.bfabric import Bfabric, BfabricAPIEngineType
+from bfabric.bfabric_config import BfabricAuth, BfabricConfig
 
-msg = "\033[93m{} version {} (2023-11-03) -- \"{}\"\
-    \nCopyright (C) 2014-2023 Functional Genomics Center Zurich\033[0m\n\n"\
-    .format(name, __version__, alias)
 
 endpoints = sorted([
         'annotation',
@@ -43,8 +40,8 @@ project = 403
 container = project
 application = 217
 
-from bfabric.bfabric import Bfabric
+
+from bfabric.bfabric_legacy import BfabricLegacy
 from bfabric.wrapper_creator.bfabric_wrapper_creator import BfabricWrapperCreator
 from bfabric.wrapper_creator.bfabric_submitter import BfabricSubmitter
 from bfabric.wrapper_creator.bfabric_feeder import BfabricFeeder
-from bfabric.bfabric_config import BfabricConfig
