+++ bfabric/scripts/bfabric_setResourceStatus_available.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 import argparse
 
-from bfabric.bfabric2 import Bfabric
+from bfabric import Bfabric
 
 
 # Copyright (C) 2014 Functional Genomics Center Zurich ETHZ|UZH. All rights reserved.
