+++ bfabric/tests/integration/test_bfabric2_exists.py
@@ -1,6 +1,7 @@
 import unittest
 
-from bfabric.bfabric2 import Bfabric, BfabricAPIEngineType, get_system_auth
+from bfabric import BfabricAPIEngineType, Bfabric
+from bfabric.bfabric import get_system_auth
 
 
 class BfabricTestExists(unittest.TestCase):
