+++ bfabric/tests/integration/test_bfabric2_read_pagination.py
@@ -1,7 +1,8 @@
 import unittest
 import pandas as pd
 
-from bfabric.bfabric2 import Bfabric, BfabricAPIEngineType, get_system_auth
+from bfabric import BfabricAPIEngineType, Bfabric
+from bfabric.bfabric import get_system_auth
 from bfabric.src.pandas_helper import list_dict_to_df
